@@ -1,14 +1,13 @@
 import React from 'react';
-import { createRoot } from 'react-dom/client';
+import ReactDOM from 'react-dom';
 import { MemoryRouter } from 'react-router-dom';
 import App from './App';
 
 it('renders without crashing', async () => {
   const div = document.createElement('div');
-  const root = createRoot(div);
-  root.render(
+  ReactDOM.render(
     <MemoryRouter>
       <App />
-    </MemoryRouter>);
+    </MemoryRouter>, div);
   await new Promise(resolve => setTimeout(resolve, 1000));
 });
