@@ -1,27 +1,30 @@
-const createProxyMiddleware = require('http-proxy-middleware');
+const { createProxyMiddleware } = require('http-proxy-middleware');
 const { env } = require('process');
 
 const target = env.ASPNETCORE_HTTPS_PORT ? `https://localhost:${env.ASPNETCORE_HTTPS_PORT}` :
-    env.ASPNETCORE_URLS ? env.ASPNETCORE_URLS.split(';')[0] : 'http://localhost:32854';
+  env.ASPNETCORE_URLS ? env.ASPNETCORE_URLS.split(';')[0] : 'http://localhost:36123';
 
 const context = [
-    "/weatherforecast",
-    "/_configuration",
-    "/.well-known",
-    "/Identity",
-    "/connect",
-    "/ApplyDatabaseMigrations",
-    "/_framework",
-
-    "/api/Entries", // �߰�
-    "/api/Books" // �߰� 
+  "/weatherforecast",
 ];
 
-module.exports = function (app) {
-    const appProxy = createProxyMiddleware(context, {
-        target: target,
-        secure: false
-    });
+const onError = (err, req, resp, target) => {
+    console.error(`${err.message}`);
+}
 
-    app.use(appProxy);
+module.exports = function (app) {
+  const appProxy = createProxyMiddleware(context, {
+    target: target,
+    // Handle errors to prevent the proxy middleware from crashing when
+    // the ASP NET Core webserver is unavailable
+    onError: onError,
+    secure: false,
+    // Uncomment this line to add support for proxying websockets
+    //ws: true, 
+    headers: {
+      Connection: 'Keep-Alive'
+    }
+  });
+
+  app.use(appProxy);
 };
