@@ -1,53 +1,13 @@
-using BookApp.Shared;
-using EntryApp.Models;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.EntityFrameworkCore;
-using RedPlus.Data;
-using RedPlus.Models;
-
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
-builder.Services.AddDbContext<ApplicationDbContext>(options =>
-    options.UseSqlServer(connectionString));
-builder.Services.AddDatabaseDeveloperPageExceptionFilter();
-
-builder.Services.AddDefaultIdentity<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = true)
-    .AddEntityFrameworkStores<ApplicationDbContext>();
-
-builder.Services.AddIdentityServer()
-    .AddApiAuthorization<ApplicationUser, ApplicationDbContext>();
-
-builder.Services.AddAuthentication()
-    .AddIdentityServerJwt();
 
 builder.Services.AddControllersWithViews();
-builder.Services.AddRazorPages();
-
-// https://github.com/dotnet/core/blob/main/release-notes/6.0/known-issues.md#aspnet-core
-builder.Services.Configure<JwtBearerOptions>(
-    "IdentityServerJwtBearer", o => o.Authority = "https://localhost:44448");
-
-#region Books, Entries 
-// EntryApp ���� ������(���Ӽ�) ���� ���� �ڵ常 ���� ��Ƽ� ���� 
-builder.Services.AddDependencyInjectionContainerForEntryApp(connectionString);
-
-// BookAppDbContext.cs Inject: New DbContext Add
-builder.Services.AddEntityFrameworkSqlServer().AddDbContext<BookAppDbContext>(options => options.UseSqlServer(connectionString));
-// IBookRepository.cs Inject: DI Container
-builder.Services.AddTransient<IBookRepository, BookRepository>();
-#endregion
 
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseMigrationsEndPoint();
-}
-else
+if (!app.Environment.IsDevelopment())
 {
     // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
@@ -57,16 +17,11 @@ app.UseHttpsRedirection();
 app.UseStaticFiles();
 app.UseRouting();
 
-app.UseAuthentication();
-app.UseIdentityServer();
-app.UseAuthorization();
 
-// app.MapDefaultControllerRoute(); // Web API 
 app.MapControllerRoute(
     name: "default",
     pattern: "{controller}/{action=Index}/{id?}");
-app.MapRazorPages();
 
-app.MapFallbackToFile("index.html"); ;
+app.MapFallbackToFile("index.html");
 
 app.Run();
