@@ -1,12 +1,13 @@
-namespace RedPlus;
-
-public class WeatherForecast
+namespace RedPlus
 {
-    public DateTime Date { get; set; }
+    public class WeatherForecast
+    {
+        public DateOnly Date { get; set; }
 
-    public int TemperatureC { get; set; }
+        public int TemperatureC { get; set; }
 
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-    public string? Summary { get; set; }
+        public string? Summary { get; set; }
+    }
 }
